+++ include/Shad/Platform.h
@@ -8,6 +8,7 @@
 #include <GL\glut.h>
 #include <string>
 #include <PolyMesh/bitmap_image.h>
+#include <Spline/stdint.h>
 
 class Platform {
 	public:
@@ -23,10 +24,10 @@ class Platform {
 		void setMoving(bool state, float deltaX, float deltaY, float deltaZ);
 		bool isMoving();
 
-		void move(int deltaPoint);
+		void move(uint64_t deltaPoint);
 
 
-		bool moveWChar(int deltaPoint, float charX, float charY, float charZ);
+		bool moveWChar(uint64_t deltaPoint, float charX, float charY, float charZ);
 
 		OpenMesh::Vec3f getDirection();
